cmd/oc-mirror: small refactoring to separate v1/v2
We shouldn't be importing the v2 module into v1 (logically wrong).
Having them as separate modules will allow us to bump their dependencies
independently from one another, enabling faster pace development in the
v2 code.

This change separates v1 and v2 executions into separate blocks. This
will make it easier to run v2 as a separate binary in the next commit.
