+++ v2/cmd/oc-mirror/main.go
@@ -3,21 +3,11 @@ package main
 import (
 	"os"
 
-	cli "github.com/openshift/oc-mirror/v2/internal/pkg/cli"
-	clog "github.com/openshift/oc-mirror/v2/internal/pkg/log"
+	cli "github.com/openshift/oc-mirror/v2/pkg/cli"
 )
 
 func main() {
-
-	// setup pluggable logger
-	// feel free to plugin you own logger
-	// just use the PluggableLoggerInterface
-	// in the file pkg/log/logger.go
-	log := clog.New("info")
-	rootCmd := cli.NewMirrorCmd(log)
-	err := rootCmd.Execute()
-	if err != nil {
-		log.Error("[Executor] %v ", err)
+	if err := cli.RunOcMirrorV2(); err != nil {
 		os.Exit(1)
 	}
 }
